+++ src/native_client/build/update_pnacl_tool_revisions.py
@@ -25,17 +25,17 @@ def ParseArgs(args):
       formatter_class=argparse.RawDescriptionHelpFormatter,
       description="""Update TOOL_REVISIONS' PNaCl version.
 
-LLVM is checked-in to the NaCl repository, but its head isn't
-necessarily the one that we currently use in PNaCl. The TOOL_REVISIONS
-file points at subversion revisions to use for tools such as LLVM. Our
-build process then downloads pre-built tool tarballs from the toolchain
-build waterfall.
+LLVM and other projects are checked-in to the NaCl repository, but their
+head isn't necessarily the one that we currently use in PNaCl. The
+TOOL_REVISIONS file points at subversion revisions to use for tools such
+as LLVM. Our build process then downloads pre-built tool tarballs from
+the toolchain build waterfall.
 
 git repository before running this script:
          ______________________
          |                    |
          v                    |
-  ...----a------b------c------d------ NaCl HEAD
+  ...----A------B------C------D------ NaCl HEAD
          ^      ^      ^      ^
          |      |      |      |__ Latest TOOL_REVISIONS update.
          |      |      |
@@ -46,29 +46,33 @@ git repository before running this script:
          |__ TOOL_REVISIONS points at an older LLVM change.
 
 git repository after running this script:
-                ______________________
-                |                    |
-                v                    |
-  ...----a------b------c------d------e------ NaCl HEAD
+                       _______________
+                       |             |
+                       v             |
+  ...----A------B------C------D------E------ NaCl HEAD
 
-Note that there could be any number of non-LLVM changes between each of
-these changelists.
-
-Performance measurements are done on the current NaCl tree, so we want
-to pull in changes under ./pnacl/ individually and let the performance
-bots measure their effect before updating to the next TOOL_REVISIONS.
+Note that there could be any number of non-PNaCl changes between each of
+these changelists, and that the user can also decide to update the
+pointer to B instead of C.
 
 There is further complication when toolchain builds are merged.
 """)
   parser.add_argument('--email', metavar='ADDRESS', type=str,
                       default=getpass.getuser()+'@chromium.org',
                       help="Email address to send errors to.")
-  parser.add_argument('--commits_since', metavar='N', type=int, default=3,
+  parser.add_argument('--commits-since', metavar='N', type=int, default=0,
                       help="Commits to wait between TOOL_REVISIONS updates.")
-  parser.add_argument('--overwrite_pnacl_svn_id', metavar='SVN_ID', type=int,
-                      help="Don't update to the oldest PNaCl change, instead "
-                      "update to the given svn id. Note that this value isn't "
-                      "checked and is assumed to be correct.")
+  parser.add_argument('--svn-id', metavar='SVN_ID', type=int, default=0,
+                      help="Update to a specific SVN ID instead of the most "
+                      "recent SVN ID with a PNaCl change. This value must "
+                      "be more recent than the one in the current "
+                      "TOOL_REVISIONS. This option is useful when multiple "
+                      "changelists' toolchain builds were merged, or when "
+                      "too many PNaCl changes would be pulled in at the "
+                      "same time.")
+  parser.add_argument('--dry-run', default=False, action='store_true',
+                      help="Print the changelist that would be sent, but "
+                      "don't actually send anything to review.")
   # TODO(jfb) The following options come from download_toolchain.py and
   #           should be shared in some way.
   parser.add_argument('--filter_out_predicates', default=[],
@@ -144,15 +148,11 @@ def GitFilesChanged(commit_hash):
                        extra=['--name-only']).split('\n')
 
 
-def GitOldestCommitChangingPath(commit_hashes, path):
-  """Returns the oldest commit changing a file under path.
-  Assumes the commit hashes are in reverse-chronological order."""
-  for commit_hash in reversed(commit_hashes):
-    files = GitFilesChanged(commit_hash)
-    for file in files:
-      if re.search('^' + path, file.strip()):
-        return commit_hash
-  return None
+def GitChangesPath(commit_hash, path):
+  """Returns True if the commit changes a file under the given path."""
+  return any([
+      re.search('^' + path, f.strip()) for f in
+      GitFilesChanged(commit_hash)])
 
 
 def GitBranchExists(name):
@@ -213,8 +213,8 @@ def CommitMessageToCleanDict(commit_message):
   fields = [
       ['git svn id',    ('\s*git-svn-id: '
                         'svn://[^@]+@([0-9]+) [a-f0-9\-]+'), '<none>'],
-      ['reviewers',     '\s*R=([^\n]+)',                   ''],
       ['reviewers tbr', '\s*TBR=([^\n]+)',                 ''],
+      ['reviewers',     '\s*R=([^\n]+)',                   ''],
       ['review url',    '\s*Review URL: *([^\n]+)',        '<none>'],
       ['bug',           '\s*BUG=([^\n]+)',                 '<none>'],
       ['test',          '\s*TEST=([^\n]+)',                '<none>'],
@@ -244,6 +244,10 @@ def SendEmail(user_email, out):
     sys.stderr.write('\nNo email address specified.')
 
 
+def DryRun(out):
+  sys.stdout.write("DRY RUN: " + out + "\n")
+
+
 def Done(out):
   sys.stdout.write(out)
   sys.exit(0)
@@ -263,8 +267,8 @@ class CLInfo:
         ('bug', None),
         ('test', None),
         ('review url', None),
-        ('reviewers', None),
         ('reviewers tbr', None),
+        ('reviewers', None),
         ('body', None),
     ])
   def __getitem__(self, key):
@@ -273,23 +277,57 @@ class CLInfo:
     assert key in self._vals.keys()
     self._vals[key] = str(val)
   def __str__(self):
-    if len([v for v in self._vals.values() if v]) == 0:
-      return ''
-    return '\n'.join([self._desc + ':'] +
-                     [('  ' + k + ':' + (' ' * (17 - len(k))) + v)
-                      for (k,v) in self._vals.items() if v]) + '\n\n'
+    """Changelist to string.
 
-def FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers, err=[], msg=[]):
+    A short description of the change, e.g.:
+      r12345: (tom@example.com) Subject of the change.
+
+    If the change is itself pulling in other changes from
+    sub-repositories then take its relevant description and append it to
+    the string. These sub-directory updates are also script-generated
+    and therefore have a predictable format. e.g.:
+      r12345: (tom@example.com) Subject of the change.
+        | dead123: (dick@example.com) Other change in another repository.
+        | beef456: (harry@example.com) Yet another cross-repository change.
+    """
+    desc = ('  r' + self._vals['git svn id'] + ': (' +
+            self._vals['author'] + ') ' +
+            self._vals['subject'])
+    if GitChangesPath(self._vals['hash'], 'pnacl/COMPONENT_REVISIONS'):
+      git_hash_abbrev = '[0-9a-fA-F]{7}'
+      email = '[^@)]+@[^)]+\.[^)]+'
+      desc = '\n'.join([desc] + [
+          '    | ' + line for line in self._vals['body'].split('\n') if
+          re.match('^ *%s: \(%s\) .*$' % (git_hash_abbrev, email), line)])
+    return desc
+
+
+def FmtOut(tr_recent, tr_points_at, pnacl_changes, err=[], msg=[]):
   assert isinstance(err, list)
   assert isinstance(msg, list)
+  old_svn_id = tr_points_at['git svn id']
+  new_svn_id = pnacl_changes[-1]['git svn id'] if pnacl_changes else '?'
+  changes = '\n'.join([str(cl) for cl in pnacl_changes])
+  bugs = '\n'.join(list(set(
+      ['BUG= ' + cl['bug'].strip() if cl['bug'] else '' for
+       cl in pnacl_changes]) - set([''])))
+  reviewers = ', '.join(list(set(
+      [r.strip() for r in
+       (','.join([
+           cl['author'] + ',' + cl['reviewers tbr'] + ',' + cl['reviewers']
+           for cl in pnacl_changes])).split(',')]) - set([''])))
   return (('*** ERROR ***\n' if err else '') +
           '\n\n'.join(err) +
           '\n\n'.join(msg) +
           ('\n\n' if err or msg else '') +
-          ('Update TOOL_REVISIONS for PNaCl %s->%s\n\n' %
-           (tr_points_at['git svn id'], oldest_pnacl['git svn id'])) +
-          str(oldest_pnacl) + str(tr_points_at) + str(tr_recent) +
-          ('\n\nR= %s\n' % reviewers if reviewers else ''))
+          ('Update TOOL_REVISIONS for PNaCl r%s->r%s\n\n'
+           'Pull the following PNaCl changes into NaCl:\n%s\n\n'
+           '%s\n'
+           'R= %s\n'
+           'TEST=git try\n'
+           'NOTRY=true\n'
+           '(Please LGTM this change and tick the "commit" box)\n' %
+           (old_svn_id, new_svn_id, changes, bugs, reviewers)))
 
 
 def Main():
@@ -298,10 +336,7 @@ def Main():
   # Updating TOOL_REVISIONS involves looking at the 3 following changelists.
   tr_recent = CLInfo('Most recent TOOL_REVISIONS update')
   tr_points_at = CLInfo('TOOL_REVISIONS update points at PNaCl version')
-  oldest_pnacl = CLInfo(
-      'Command-line specified PNaCl change' if args.overwrite_pnacl_svn_id
-      else 'Oldest PNaCl change since this PNaCl version')
-  reviewers = None
+  pnacl_changes = []
   msg = []
 
   try:
@@ -318,22 +353,21 @@ def Main():
       tr_recent[i] = GitCommitInfo(info=i, obj='TOOL_REVISIONS', num=1)
     recent_commits = GitCommitsSince(tr_recent['date'])
     tr_recent['commits since'] = len(recent_commits)
-    assert len(recent_commits) > 1
     if len(recent_commits) - 1 < args.commits_since:
-      # This check is concervative: the TOOL_REVISIONS update might not
+      # This check is conservative: the TOOL_REVISIONS update might not
       # have changed PNACL_VERSION. We might wait longer than we
       # actually need to.
       # TODO(jfb) Look at the previous commit, and if args.commit_since is
       #           respected (none of the previous commits actually change
       #           PNACL_VERSION) then do the update.
-      Done(FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers,
+      Done(FmtOut(tr_recent, tr_points_at, pnacl_changes,
                   msg=['Nothing to do: not enough commits since last '
                        'TOOL_REVISIONS update, need at least %s commits to '
                        'ensure that we get enough performance runs.' %
                        args.commits_since]))
 
     # The current TOOL_REVISIONS points at a specific PNaCl LLVM
-    # version.  LLVM is checked-in to the NaCl repository, but its head
+    # version. LLVM is checked-in to the NaCl repository, but its head
     # isn't necessarily the one that we currently use in PNaCl.
     current_versions = download_toolchains.LoadVersions('TOOL_REVISIONS')
     tr_points_at['git svn id'] = current_versions['PNACL_VERSION']
@@ -346,59 +380,60 @@ def Main():
     tr_points_at['commits since'] = len(recent_commits)
     assert len(recent_commits) > 1
 
-    if args.overwrite_pnacl_svn_id:
-      # The user doesn't want to update TOOL_REVISIONS to point at the
-      # oldest PNaCl change that follows the previous TOOL_REVISIONS
-      # pointer.  This could happens when it makes sense to skip
-      # innocuous changes, or when moving the pointer backwards which
-      # only a user is qualified to decide to do.
-      oldest_pnacl['hash'] = FindCommitWithGitSvnId(
-          str(args.overwrite_pnacl_svn_id))
-      if not oldest_pnacl['hash']:
-        Done(FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers,
-                    err=[('Could not find git svn id %s provided at '
-                          'the command line.') %
-                         (args.overwrite_pnacl_svn_id)]))
-    else:
-      # Find the oldest PNaCl change that follows the previous
-      # TOOL_REVISIONS PNaCl pointer. The default only updates to the
-      # next such change to isolate each LLVM commit: it simplifies bug
-      # and regression finding.
-      oldest_pnacl['hash'] = GitOldestCommitChangingPath(
-          recent_commits[:-1], 'pnacl/')
-      if not oldest_pnacl['hash']:
-        Done(FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers,
-                    msg=['No PNaCl change since.']))
+    if args.svn_id and args.svn_id <= int(tr_points_at['git svn id']):
+      Done(FmtOut(tr_recent, tr_points_at, pnacl_changes,
+                  err=["Can't update to SVN ID r%s, the current "
+                       "TOOL_REVISIONS' SVN ID (r%s) is more recent." %
+                       (args.svn_id, tr_points_at['git svn id'])]))
 
-    for i in ['author', 'date', 'subject']:
-      oldest_pnacl[i] = GitCommitInfo(info=i, obj=oldest_pnacl['hash'], num=1)
-    for k,v in CommitMessageToCleanDict(
-        GitCommitInfo(info='body',
-                        obj=oldest_pnacl['hash'], num=1)).iteritems():
-      oldest_pnacl[k] = v
-    oldest_pnacl['body'] = (
-        '\n  ' + '='*76 + '\n  ' +
-        oldest_pnacl['body'].replace('\n', '\n  ') +
-        '\n  ' + '='*76)
-    reviewers = ', '.join(set(
-        [r.strip() for r in ([oldest_pnacl['author']] +
-                             oldest_pnacl['reviewers'].split(',') +
-                             oldest_pnacl['reviewers tbr'].split(',')
-                             ) if r.strip()]))
+    # Find the commits changing PNaCl files that follow the previous
+    # TOOL_REVISIONS PNaCl pointer.
+    pnacl_pathes = ['pnacl/', 'toolchain_build/']
+    pnacl_hashes = list(set(reduce(
+        lambda acc, lst: acc + lst,
+        [[cl for cl in recent_commits[:-1] if
+          GitChangesPath(cl, path)] for
+         path in pnacl_pathes])))
+    for hash in pnacl_hashes:
+      cl = CLInfo('PNaCl change ' + hash)
+      cl['hash'] = hash
+      for i in ['author', 'date', 'subject']:
+        cl[i] = GitCommitInfo(info=i, obj=hash, num=1)
+      for k,v in CommitMessageToCleanDict(
+          GitCommitInfo(info='body', obj=hash, num=1)).iteritems():
+        cl[k] = v
+      pnacl_changes.append(cl)
+
+    # The PNaCl hashes weren't ordered chronologically, make sure the
+    # changes are.
+    pnacl_changes.sort(key=lambda x: int(x['git svn id']))
+
+    if args.svn_id:
+      pnacl_changes = [cl for cl in pnacl_changes if
+                       int(cl['git svn id']) <= args.svn_id]
+
+    if len(pnacl_changes) == 0:
+      Done(FmtOut(tr_recent, tr_points_at, pnacl_changes,
+                  msg=['No PNaCl change since r%s.' %
+                       tr_points_at['git svn id']]))
+
+    current_versions['PNACL_VERSION'] = pnacl_changes[-1]['git svn id']
 
     new_branch_name = ('TOOL_REVISIONS-pnacl-update-to-%s' %
-                       oldest_pnacl['git svn id'])
+                       current_versions['PNACL_VERSION'])
     if GitBranchExists(new_branch_name):
       # TODO(jfb) Figure out if git-try succeeded, checkout the branch
       #           and dcommit.
       raise Exception("Branch %s already exists, the change hasn't "
                       "landed yet.\nPlease check trybots and dcommit it "
                       "manually." % new_branch_name)
-    GitCheckoutNewBranch(new_branch_name)
+    if args.dry_run:
+      DryRun("Would check out branch: " + new_branch_name)
+    else:
+      GitCheckoutNewBranch(new_branch_name)
 
     # Change the value of PNACL_VERSION in the TOOL_REVISIONS file to
     # the desired git svn id. The hashes are updated below.
-    current_versions['PNACL_VERSION'] = oldest_pnacl['git svn id']
     updates = [['PNACL_VERSION',
                 tr_points_at['git svn id'],
                 current_versions['PNACL_VERSION']]]
@@ -415,8 +450,9 @@ def Main():
         old_value = current_versions[keyname]
         updates.append([keyname, old_value, new_value])
     except urllib2.HTTPError as e:
-      GitCheckout('master', force=True)
-      GitDeleteBranch(new_branch_name, force=True)
+      if not args.dry_run:
+        GitCheckout('master', force=True)
+        GitDeleteBranch(new_branch_name, force=True)
       # TODO(jfb) Automatically split the commits.
       #           Before doing this, we should try the next versions and see
       #           if they have a toolchain and don't touch PNaCl.
@@ -430,31 +466,35 @@ def Main():
           '(merged toolchain build).\n'
           '  Force build with your name, reason, no branch, '
           'and revision %s.' %
-          (e.url, e.code, e.reason, oldest_pnacl['git svn id']))
+          (e.url, e.code, e.reason, current_versions['PNACL_VERSION']))
 
-    # Change the TOOL_REVISIONS file in the local git repository.
-    with open('TOOL_REVISIONS', 'r+') as tool_revisions:
-      content = tool_revisions.read()
-      for keyname, old_value, new_value in updates:
-        old = '%s=%s' % (keyname, old_value)
-        new = '%s=%s' % (keyname, new_value)
-        assert content.find(old) != -1, ("TOOL_REVISIONS doesn't contain %s" %
-                                         old)
-        content = content.replace(old, new)
-      tool_revisions.seek(0)
-      tool_revisions.truncate()
-      tool_revisions.write(content)
-    GitAdd('TOOL_REVISIONS')
-    GitCommit(FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers))
+    if args.dry_run:
+      DryRun("Would update TOOL_REVISIONS to:\n%s" %
+             '\n'.join(['  ' + u[0] + '=' + u[2] for u in updates]))
+    else:
+      # Change the TOOL_REVISIONS file in the local git repository.
+      with open('TOOL_REVISIONS', 'r+') as tool_revisions:
+        content = tool_revisions.read()
+        for keyname, old_value, new_value in updates:
+          old = '%s=%s' % (keyname, old_value)
+          new = '%s=%s' % (keyname, new_value)
+          assert content.find(old) != -1, (
+              "TOOL_REVISIONS doesn't contain %s" % old)
+          content = content.replace(old, new)
+        tool_revisions.seek(0)
+        tool_revisions.truncate()
+        tool_revisions.write(content)
+      GitAdd('TOOL_REVISIONS')
+      GitCommit(FmtOut(tr_recent, tr_points_at, pnacl_changes))
 
-    upload_res = UploadChanges()
-    msg += ['Upload result:\n%s' % upload_res]
-    try_res = GitTry()
-    msg += ['Try result:\n%s' % try_res]
+      upload_res = UploadChanges()
+      msg += ['Upload result:\n%s' % upload_res]
+      try_res = GitTry()
+      msg += ['Try result:\n%s' % try_res]
 
-    GitCheckout('master', force=False)
+      GitCheckout('master', force=False)
 
-    Done(FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers, msg=msg))
+    Done(FmtOut(tr_recent, tr_points_at, pnacl_changes, msg=msg))
 
   except SystemExit as e:
     # Normal exit.
@@ -463,11 +503,12 @@ def Main():
   except (BaseException, Exception) as e:
     # Leave the branch around, if any was created: it'll prevent next
     # runs of the cronjob from succeeding until the failure is fixed.
-    out = FmtOut(tr_recent, tr_points_at, oldest_pnacl, reviewers, msg=msg,
+    out = FmtOut(tr_recent, tr_points_at, pnacl_changes, msg=msg,
                  err=['Failed at %s: %s' % (datetime.datetime.now(), e)])
     sys.stderr.write(out)
-    SendEmail(args.email, out)
-    GitCheckout('master', force=True)
+    if not args.dry_run:
+      SendEmail(args.email, out)
+      GitCheckout('master', force=True)
     raise
